+++ src/data/protocols/db/breed/index.ts
@@ -1,0 +1,2 @@
+export * from './load-cat-breeds-repository'
+export * from './load-dog-breeds-repository'
