+++ src/data/protocols/db/index.ts
@@ -2,3 +2,4 @@ export * from './guardian'
 export * from './log'
 export * from './pet'
 export * from './specie'
+export * from './breed'
