+++ src/domain/use-cases/pet/breed/load-dog-breeds.ts
@@ -1,4 +1,4 @@
-import { type LoadDogBreedsRepository } from '@/data/protocols/db/breed/load-dog-breeds-repository'
+import { type LoadDogBreedsRepository } from '@/data/protocols/db/breed'
 import { type Breed } from '@/domain/models/breed'
 
 export interface LoadDogBreeds {
