+++ src/domain/use-cases/pet/index.ts
@@ -1,2 +1,3 @@
 export * from './add-pet'
 export * from './appoint-specie'
+export * from './breed'
